PopularPerson Fragment Added

+++ app/src/main/java/com/michaelkatan/moviedatabaseapp/models/Person.kt
@@ -1,4 +1,4 @@
 package com.michaelkatan.moviedatabaseapp.models
 
 data class Person(val id: Int, val profile_path: String, val adult: Boolean,
-                    val name: String, val popularity: Int)
+                    val name: String, val popularity: Double)
